refactor(compiler): rename TaggingCompiler into Compiler and TaggingViewFactory into ViewFactory
Remove Compiler and ViewFactory interfaces because they have only one implementation each.
