+++ lib/core_dom/module_internal.dart
@@ -43,8 +43,6 @@ part 'node_cursor.dart';
 part 'selector.dart';
 part 'shadow_dom_component_factory.dart';
 part 'shadowless_shadow_root.dart';
-part 'tagging_compiler.dart';
-part 'tagging_view_factory.dart';
 part 'template_cache.dart';
 part 'transcluding_component_factory.dart';
 part 'tree_sanitizer.dart';
@@ -68,7 +66,7 @@ class CoreDomModule extends Module {
     bind(TextMustache);
     bind(AttrMustache);
 
-    bind(Compiler, toImplementation: TaggingCompiler);
+    bind(Compiler);
     bind(CompilerConfig);
 
     bind(ComponentFactory, toInstanceOf: SHADOW_DOM_COMPONENT_FACTORY_KEY);
